2-ecs-microservice-stack/requirements.txt
aws_cdk.core
aws_cdk.aws_ec2
aws_cdk.aws_ecs
aws_cdk.aws_rds
aws_cdk.aws_ecs_patterns
